requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.10.3
cryptography==45.0.6
defusedxml==0.7.1
Django==5.2.4
django-appconf==1.1.0
django-baton==5.0.2
django-imagekit==5.0.0
django-oauth-toolkit==3.0.1
django-rest-framework-social-oauth2==1.2.0
django-rest-passwordreset==1.5.0
djangorestframework==3.16.0
dnspython==2.7.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
eventlet==0.40.2
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
kombu==5.5.4
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pilkit==3.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
responses==0.25.7
rpds-py==0.27.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
zope.event==5.1.1
zope.interface==7.2
